move old team page contents to compare teams page as a starting point
